Created sound list viewmodel to reduce code duplication
