+++ src/AmbientSounds.Uwp/AmbientSounds.Uwp.csproj
@@ -136,6 +136,9 @@
     <Compile Include="Controls\SleepTimerControl.xaml.cs">
       <DependentUpon>SleepTimerControl.xaml</DependentUpon>
     </Compile>
+    <Compile Include="Controls\SoundGridControl.xaml.cs">
+      <DependentUpon>SoundGridControl.xaml</DependentUpon>
+    </Compile>
     <Compile Include="Converters\LocalizationConverter.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Services\LocalSettings.cs" />
@@ -253,6 +256,10 @@
       <Generator>MSBuild:Compile</Generator>
     </Page>
     <Page Include="Controls\SleepTimerControl.xaml">
+      <SubType>Designer</SubType>
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
+    <Page Include="Controls\SoundGridControl.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
