+++ src/AmbientSounds.Uwp/Controls/SoundGridControl.xaml
@@ -1,0 +1,109 @@
+﻿<UserControl
+    x:Class="AmbientSounds.Controls.SoundGridControl"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:animations="using:Microsoft.Toolkit.Uwp.UI.Animations"
+    xmlns:converters="using:AmbientSounds.Converters"
+    xmlns:core="using:Microsoft.Xaml.Interactions.Core"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:extensions="using:Microsoft.Toolkit.Uwp.UI.Extensions"
+    xmlns:interactivity="using:Microsoft.Xaml.Interactivity"
+    xmlns:local="using:AmbientSounds.Controls"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    xmlns:muxc="using:Microsoft.UI.Xaml.Controls"
+    xmlns:viewmodels="using:AmbientSounds.ViewModels"
+    d:DesignHeight="300"
+    d:DesignWidth="400"
+    mc:Ignorable="d">
+    <interactivity:Interaction.Behaviors>
+        <core:EventTriggerBehavior EventName="Loaded">
+            <core:InvokeCommandAction Command="{x:Bind ViewModel.LoadCommand}" />
+        </core:EventTriggerBehavior>
+    </interactivity:Interaction.Behaviors>
+
+    <Grid>
+        <ScrollViewer Padding="40,32,40,0" CanContentRenderOutsideBounds="True">
+            <GridView
+                animations:ReorderGridAnimation.Duration="300"
+                AutomationProperties.Name="Sounds"
+                IsItemClickEnabled="True"
+                ItemContainerStyle="{StaticResource GridViewItemExpanded}"
+                ItemsSource="{x:Bind ViewModel.Sounds}"
+                ScrollViewer.HorizontalScrollMode="Disabled"
+                ScrollViewer.VerticalScrollMode="Disabled"
+                SelectionMode="None">
+                <TransitionCollection>
+                    <EntranceThemeTransition IsStaggeringEnabled="True" />
+                </TransitionCollection>
+                <interactivity:Interaction.Behaviors>
+                    <core:ItemClickBehavior Command="{x:Bind ViewModel.PlaySoundCommand}" />
+                </interactivity:Interaction.Behaviors>
+                <GridView.ItemTemplate>
+                    <DataTemplate x:DataType="viewmodels:SoundViewModel">
+                        <GridViewItem
+                            Width="300"
+                            Height="180"
+                            AutomationProperties.Name="{x:Bind converters:LocalizationConverter.ConvertGridViewSoundName(IsCurrentlyPlaying, Unplayable, Name), Mode=OneWay}">
+                            <GridViewItem.Resources>
+                                <!--  These two removes the border on hover and on pressed  -->
+                                <SolidColorBrush x:Key="SystemControlHighlightListLowBrush" Color="Transparent" />
+                                <SolidColorBrush x:Key="SystemControlHighlightListMediumBrush" Color="Transparent" />
+                            </GridViewItem.Resources>
+                            <Grid
+                                Width="280"
+                                Height="160"
+                                Margin="4"
+                                extensions:VisualExtensions.CenterPoint="140,80,0"
+                                CornerRadius="8"
+                                PointerEntered="GridScaleUp"
+                                PointerExited="GridScaleNormal">
+                                <animations:Implicit.Animations>
+                                    <animations:ScaleAnimation Duration="0:0:0.5" />
+                                </animations:Implicit.Animations>
+                                <Grid.Background>
+                                    <ImageBrush Stretch="UniformToFill">
+                                        <ImageBrush.ImageSource>
+                                            <BitmapImage UriSource="{x:Bind ImagePath}" />
+                                        </ImageBrush.ImageSource>
+                                    </ImageBrush>
+                                </Grid.Background>
+                                <Grid VerticalAlignment="Bottom">
+                                    <Grid.Background>
+                                        <AcrylicBrush TintColor="{ThemeResource SystemChromeMediumLowColor}" TintOpacity="0.2" />
+                                    </Grid.Background>
+                                    <TextBlock
+                                        Margin="12,0,0,4"
+                                        FontFamily="Calibri"
+                                        FontSize="36"
+                                        FontWeight="SemiBold"
+                                        Foreground="White"
+                                        Text="{x:Bind converters:LocalizationConverter.ConvertSoundName(Name)}"
+                                        TextWrapping="Wrap" />
+                                </Grid>
+                            </Grid>
+                            <GridViewItem.ContextFlyout>
+                                <Flyout Placement="BottomEdgeAlignedLeft">
+                                    <StackPanel Width="300">
+                                        <TextBlock x:Uid="Copyright" FontWeight="SemiBold" />
+                                        <TextBlock
+                                            Margin="0,8,0,0"
+                                            IsTextSelectionEnabled="True"
+                                            Text="{x:Bind Attribution}"
+                                            TextWrapping="Wrap" />
+                                    </StackPanel>
+                                </Flyout>
+                            </GridViewItem.ContextFlyout>
+                        </GridViewItem>
+                    </DataTemplate>
+                </GridView.ItemTemplate>
+            </GridView>
+        </ScrollViewer>
+
+        <muxc:ProgressRing
+            Width="40"
+            Height="40"
+            HorizontalAlignment="Center"
+            VerticalAlignment="Center"
+            IsActive="{x:Bind ViewModel.LoadCommand.IsRunning, Mode=OneWay}" />
+    </Grid>
+</UserControl>
