+++ src/AmbientSounds.Uwp/Views/MainPage.xaml
@@ -2,24 +2,12 @@
     x:Class="AmbientSounds.Views.MainPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:animations="using:Microsoft.Toolkit.Uwp.UI.Animations"
     xmlns:controls="using:AmbientSounds.Controls"
-    xmlns:converters="using:AmbientSounds.Converters"
-    xmlns:core="using:Microsoft.Xaml.Interactions.Core"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:extensions="using:Microsoft.Toolkit.Uwp.UI.Extensions"
-    xmlns:interactivity="using:Microsoft.Xaml.Interactivity"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:muxc="using:Microsoft.UI.Xaml.Controls"
     xmlns:statetriggers="using:AmbientSounds.StateTriggers"
-    xmlns:viewmodels="using:AmbientSounds.ViewModels"
     Background="{ThemeResource ApplicationPageBackgroundThemeBrush}"
     mc:Ignorable="d">
-    <interactivity:Interaction.Behaviors>
-        <core:EventTriggerBehavior EventName="Loaded">
-            <core:InvokeCommandAction Command="{x:Bind ViewModel.LoadCommand}" />
-        </core:EventTriggerBehavior>
-    </interactivity:Interaction.Behaviors>
 
     <Grid>
         <Grid.RowDefinitions>
@@ -55,94 +43,7 @@
             </StackPanel>
         </Grid>
 
-        <ScrollViewer
-            x:Name="SoundsViewer"
-            Grid.Row="1"
-            Padding="40,32,40,0"
-            CanContentRenderOutsideBounds="True">
-            <GridView
-                animations:ReorderGridAnimation.Duration="300"
-                AutomationProperties.Name="Sounds"
-                IsItemClickEnabled="True"
-                ItemContainerStyle="{StaticResource GridViewItemExpanded}"
-                ItemsSource="{x:Bind ViewModel.Sounds}"
-                ScrollViewer.HorizontalScrollMode="Disabled"
-                ScrollViewer.VerticalScrollMode="Disabled"
-                SelectionMode="None">
-                <TransitionCollection>
-                    <EntranceThemeTransition IsStaggeringEnabled="True" />
-                </TransitionCollection>
-                <interactivity:Interaction.Behaviors>
-                    <core:ItemClickBehavior Command="{x:Bind ViewModel.PlaySoundCommand}" />
-                </interactivity:Interaction.Behaviors>
-                <GridView.ItemTemplate>
-                    <DataTemplate x:DataType="viewmodels:SoundViewModel">
-                        <GridViewItem
-                            Width="300"
-                            Height="180"
-                            AutomationProperties.Name="{x:Bind converters:LocalizationConverter.ConvertGridViewSoundName(IsCurrentlyPlaying, Unplayable, Name), Mode=OneWay}">
-                            <GridViewItem.Resources>
-                                <!--  These two removes the border on hover and on pressed  -->
-                                <SolidColorBrush x:Key="SystemControlHighlightListLowBrush" Color="Transparent" />
-                                <SolidColorBrush x:Key="SystemControlHighlightListMediumBrush" Color="Transparent" />
-                            </GridViewItem.Resources>
-                            <Grid
-                                Width="280"
-                                Height="160"
-                                Margin="4"
-                                extensions:VisualExtensions.CenterPoint="140,80,0"
-                                CornerRadius="8"
-                                PointerEntered="GridScaleUp"
-                                PointerExited="GridScaleNormal">
-                                <animations:Implicit.Animations>
-                                    <animations:ScaleAnimation Duration="0:0:0.5" />
-                                </animations:Implicit.Animations>
-                                <Grid.Background>
-                                    <ImageBrush Stretch="UniformToFill">
-                                        <ImageBrush.ImageSource>
-                                            <BitmapImage UriSource="{x:Bind ImagePath}" />
-                                        </ImageBrush.ImageSource>
-                                    </ImageBrush>
-                                </Grid.Background>
-                                <Grid VerticalAlignment="Bottom">
-                                    <Grid.Background>
-                                        <AcrylicBrush TintColor="{ThemeResource SystemChromeMediumLowColor}" TintOpacity="0.2" />
-                                    </Grid.Background>
-                                    <TextBlock
-                                        Margin="12,0,0,4"
-                                        FontFamily="Calibri"
-                                        FontSize="36"
-                                        FontWeight="SemiBold"
-                                        Foreground="White"
-                                        Text="{x:Bind converters:LocalizationConverter.ConvertSoundName(Name)}"
-                                        TextWrapping="Wrap" />
-                                </Grid>
-                            </Grid>
-                            <GridViewItem.ContextFlyout>
-                                <Flyout Placement="BottomEdgeAlignedLeft">
-                                    <StackPanel Width="300">
-                                        <TextBlock x:Uid="Copyright" FontWeight="SemiBold" />
-                                        <TextBlock
-                                            Margin="0,8,0,0"
-                                            IsTextSelectionEnabled="True"
-                                            Text="{x:Bind Attribution}"
-                                            TextWrapping="Wrap" />
-                                    </StackPanel>
-                                </Flyout>
-                            </GridViewItem.ContextFlyout>
-                        </GridViewItem>
-                    </DataTemplate>
-                </GridView.ItemTemplate>
-            </GridView>
-        </ScrollViewer>
-
-        <muxc:ProgressRing
-            Grid.Row="1"
-            Width="40"
-            Height="40"
-            HorizontalAlignment="Center"
-            VerticalAlignment="Center"
-            IsActive="{x:Bind ViewModel.LoadCommand.IsRunning, Mode=OneWay}" />
+        <controls:SoundGridControl x:Name="SoundsViewer" Grid.Row="1" />
 
         <VisualStateManager.VisualStateGroups>
             <VisualStateGroup>
@@ -162,7 +63,7 @@
                     <VisualState.Setters>
                         <Setter Target="TitleCentre.Visibility" Value="Visible" />
                         <Setter Target="HeaderGrid.Padding" Value="88,55,88,28" />
-                        <Setter Target="SoundsViewer.Padding" Value="88,32,88,0" />
+                        <Setter Target="SoundsViewer.Margin" Value="48,0,48,0" />
                     </VisualState.Setters>
                 </VisualState>
             </VisualStateGroup>
