+++ src/AmbientSounds.Uwp/Views/MainPage.xaml.cs
@@ -1,10 +1,5 @@
-﻿using AmbientSounds.ViewModels;
-using Microsoft.Extensions.DependencyInjection;
-using System.Numerics;
-using Windows.UI.ViewManagement;
-using Windows.UI.Xaml;
+﻿using Windows.UI.ViewManagement;
 using Windows.UI.Xaml.Controls;
-using Windows.UI.Xaml.Hosting;
 
 namespace AmbientSounds.Views
 {
@@ -16,31 +11,12 @@ namespace AmbientSounds.Views
         public MainPage()
         {
             this.InitializeComponent();
-            this.DataContext = App.Services.GetRequiredService<MainPageViewModel>();
 
             if (App.IsTenFoot)
             {
                 // Ref: https://docs.microsoft.com/en-us/windows/uwp/xbox-apps/turn-off-overscan
                 ApplicationView.GetForCurrentView().SetDesiredBoundsMode(ApplicationViewBoundsMode.UseCoreWindow);
             }
-        }
-
-        public MainPageViewModel ViewModel => (MainPageViewModel)this.DataContext;
-
-        private void GridScaleUp(object sender, Windows.UI.Xaml.Input.PointerRoutedEventArgs e)
-        {
-            // Source for the scaling: https://github.com/windows-toolkit/WindowsCommunityToolkit/blob/master/Microsoft.Toolkit.Uwp.SampleApp/SamplePages/Implicit%20Animations/ImplicitAnimationsPage.xaml.cs
-            // Search for "Scale Element".
-            var element = sender as UIElement;
-            var visual = ElementCompositionPreview.GetElementVisual(element);
-            visual.Scale = new Vector3(1.1f, 1.1f, 1);
-        }
-
-        private void GridScaleNormal(object sender, Windows.UI.Xaml.Input.PointerRoutedEventArgs e)
-        {
-            var element = sender as UIElement;
-            var visual = ElementCompositionPreview.GetElementVisual(element);
-            visual.Scale = new Vector3(1);
         }
     }
 }
